pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.123.9",
    "html5lib>=1.1",
    "jinja2>=3.1.6",
    "lxml>=5.0.0",
    "pandas>=2.3.3",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.0",
    "ruff>=0.14.8",
]
